refactored group components
