+++ app/(authenticated)/user/groups/[id]/page.tsx
@@ -12,7 +12,7 @@ import { useAvatar } from "@/hooks/useAvatar"
 import Image from "next/image"
 import { CreateLaagDialog } from "./create-laag-dialog"
 import { LaagFeed } from "./laag-feed"
-import { EditGroupModal } from "@/components/edit-group-modal"
+import { EditGroupModal } from "@/components/groups/edit-group-modal"
 import { ManageMembersDialog } from "./manage-members-dialog"
 import { Button } from "@/components/ui/button"
 import { useAuth } from "@/app/context/auth-context"
