+++ components/app/app-sidebar.tsx
@@ -4,11 +4,9 @@
 import * as React from "react"
 import {
   AudioWaveform,
-  Baby,
   BookOpen,
   Bot,
   Command,
-  Frame,
   GalleryVerticalEnd,
   Map,
   PieChart,
@@ -17,10 +15,10 @@ import {
   UsersRound,
 } from "lucide-react"
 
-import { NavMain } from "@/components/nav-main"
-import { NavProjects } from "@/components/nav-projects"
-import { NavUser } from "@/components/nav-user"
-import { TeamSwitcher } from "@/components/team-switcher"
+import { NavMain } from "@/components/app/nav-main"
+import { NavProjects } from "@/components/app/nav-projects"
+import { NavUser } from "@/components/app/nav-user"
+import { TeamSwitcher } from "@/components/app/team-switcher"
 import {
   Sidebar,
   SidebarContent,
